+++ .lintstagedrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '*.php': 'vendor/bin/phpcs --standard=phpcs.xml.dist',
+};